Move session-wide env from bashrc to profile
If i3 is launched from a display manager, bash is not the parent process
and env variables like TERMINAL will not be setup. This results in a
malfunctioning i3-sensible-terminal, if you have a x-terminal-emulator
symlink (as on Ubuntu).

This commit moves TERMINAL and some other vars to .profile so they get
sourced by the display manager.