ssh: add aliases for reverse-proxying
Adds an alias for setting up a SSH reverse-proxy through revanna.

This is useful if you're firewalled, NAT'd and without VPN but want
to reach your machine from the outside via ssh.

    firewalled$ rrproxy

And then just ssh to "rp" from the outside and it'll ProxyJump via
revanna back into the host using the outbound ssh-connection.