Include YADM 2.4 in dotfiles
Since not all distros ship YADM 2.x yet (Ubuntu), include a copy in
~/bin/ to make sure we run the same version across all machines.
You could achieve the same thing with a Git submodule, but that doesn't
work well if you're in an environment without Internet-access often.
This allows the entire dotfiles repo to be distributed as a tarball, if
need be.

The bundled YADM is v2.4.